refactoring

+++ src/main/java/com/fnb/front/backend/service/CartService.java
@@ -42,8 +42,13 @@ public class CartService {
     }
 
     @Transactional
-    public boolean delete(int cartId) {
-        this.cartRepository.deleteCart(cartId);
+    public boolean delete(int cartId, String memberId) {
+        int count = this.cartRepository.deleteCart(cartId, memberId);
+
+        if (count == 0) {
+            throw new IllegalArgumentException("장바구니가 존재하지 않습니다.");
+        }
+
         this.cartRepository.deleteCartItem(cartId);
 
         return true;
@@ -52,8 +57,6 @@ public class CartService {
     public boolean update(CartUpdateRequest cartUpdateRequest) {
         Cart cart = this.cartRepository.findCart(cartUpdateRequest.getCartId());
 
-        //exception
-        // 동시적 요청으로 인해 존재가 확인되지 않음은 exception 처리
         if (cart == null) {
             throw new IllegalArgumentException("장바구니가 존재하지 않습니다.");
         }
